requirements.txt
SQLAlchemy
envparse
pydantic
requests
python-telegram-bot
psycopg2-binary
pytz
grpcio-tools
protobuf
gcloud
google-api-python-client
python-telegram-bot-pagination
aiohttp
beautifulsoup4
Pillow
gtts